Agrega validaciones y mejoras en servicios de clientes

+++ src/main/java/py/com/ecommerce/cliente/service/ClienteServiceImpl.java
@@ -1,5 +1,6 @@
 package py.com.ecommerce.cliente.service;
 
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,7 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 @Slf4j
+@Transactional
 public class ClienteServiceImpl implements ClienteService{
     private final ClienteRepository clienteRepository;
 
@@ -40,17 +42,39 @@ public class ClienteServiceImpl implements ClienteService{
     @Override
     public List<ClienteEntity> getAll() {
         log.info("Procedemos a obtener todos los clientes de la base de datos");
-        return clienteRepository.findAll();
+        List<ClienteEntity> clientes = clienteRepository.findAll();
+
+        if (clientes.isEmpty()) {
+            log.warn("La base de datos no contiene clientes.");
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    "No existen clientes en la base de datos"
+            );
+        }
+        log.info("Se obtuvo la lista de clientes con exito");
+        return clientes;
     }
 
-    @Override
-    public ClienteEntity saveOrUpdate(ClienteEntity cliente) {
-        log.info("Procedemos a persistir el cliente {}", cliente.getIdCliente());
+    public ClienteEntity createCliente(ClienteEntity cliente) {
+        if (cliente.getIdCliente() != null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No se debe enviar ID para crear un cliente.");
+        }
+        log.info("Procedemos a guardar el cliente nuevo");
+        return clienteRepository.save(cliente);
+    }
+
+
+    public ClienteEntity updateCliente(Long id, ClienteEntity clienteRequest) {
+        ClienteEntity cliente = clienteRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Cliente con ID " + id + " no existe para actualizar."));
+        log.info("Procedemos a actualizar el cliente {}", cliente.getIdCliente());
+        cliente.setIdCliente(id);
         return clienteRepository.save(cliente);
     }
 
     @Override
     public EliminarClienteRespuestaDto delete(Long id) {
+        if (!clienteRepository.existsById(id))  throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El cliente con ID " + id + " no existe.");
         try {
             log.info("Procedemos a eliminar el cliente con id {}", id);
             clienteRepository.deleteById(id);
